chore: add required co-dependencies
