+++ packages/aio-commerce-lib-auth/source/index.ts
@@ -14,4 +14,3 @@ export * from "./lib/ims-auth";
 export * from "./lib/ims-auth/ims-auth-types";
 export * from "./lib/integration-auth";
 export * from "./lib/integration-auth/integration-auth-types";
-export * from "./lib/validation";
