+++ packages/aio-commerce-lib-core/package.json
@@ -3,26 +3,21 @@
   "type": "module",
   "version": "1.0.0",
   "private": false,
-
   "author": "Adobe Inc.",
   "engines": {
     "node": ">=22"
   },
-
   "license": "Apache-2.0",
   "description": "Description for @adobe/aio-commerce-lib-core",
   "keywords": [],
-
   "bugs": {
     "url": "https://github.com/adobe/aio-commerce-sdk/issues"
   },
-
   "repository": {
     "type": "git",
     "url": "https://github.com/adobe/aio-commerce-sdk.git",
     "directory": "packages/aio-commerce-lib-core"
   },
-
   "main": "./dist/cjs/index.cjs",
   "module": "./dist/es/index.js",
   "types": "./dist/es/index.d.ts",
@@ -39,7 +34,6 @@
     },
     "./package.json": "./package.json"
   },
-
   "scripts": {
     "build": "tsdown",
     "check:ci": "pnpm -w check:ci .",
@@ -51,13 +45,14 @@
     "test": "vitest run",
     "test:watch": "vitest --watch"
   },
-
-  "dependencies": {},
+  "dependencies": {
+    "ansis": "^4.1.0",
+    "valibot": "^1.1.0"
+  },
   "devDependencies": {
-    "@aio-commerce-sdk/config-typescript": "workspace:*",
     "@aio-commerce-sdk/config-tsdown": "workspace:*",
+    "@aio-commerce-sdk/config-typescript": "workspace:*",
     "vitest": "^2.0.0"
   },
-
   "sideEffects": false
 }
