+++ packages/aio-commerce-lib-core/source/index.ts
@@ -1,1 +1,2 @@
 export * from "./lib/result";
+export * from "./lib/validation";
