--- packages/aio-commerce-lib-core/test/index.test.ts
@@ -1,7 +1,0 @@
-import { describe, expect, it } from "vitest";
-
-describe("aio-commerce-lib-core", () => {
-  it("should work", () => {
-    expect(true).toBe(true);
-  });
-});
