+++ packages/aio-commerce-sdk/source/index.ts
@@ -1,1 +1,2 @@
 export * as Auth from "@adobe/aio-commerce-lib-auth";
+export * as Core from "@adobe/aio-commerce-lib-core";
