+++ turbo.json
@@ -13,7 +13,7 @@
       "dependsOn": ["^lint"]
     },
     "test": {
-      "dependsOn": ["^test"]
+      "dependsOn": ["^build", "^test"]
     },
     "typecheck": {
       "dependsOn": ["^build", "^typecheck"]
